after refactor of CountryDaoImpl
